회원 가입 체육관 리스트 개편 / 전화번호 스타일 파셜화 / 회원가입시 추천인 코드 입력하는 기능 일부 구현

+++ db/migrate/20200713051026_drop_line_items.rb
@@ -1,5 +1,7 @@
 class DropLineItems < ActiveRecord::Migration[6.0]
   def change
-    drop_table :line_items
+    if table_exists?(:line_items)
+      drop_table :line_items
+    end
   end
 end
